[TimePicker]: Refactor to BaseTimePicker class to re-use functionality
