+++ uui-components/src/inputs/BaseTimePicker.tsx
@@ -1,0 +1,91 @@
+import * as React from 'react';
+import dayjs from 'dayjs';
+import { isChildFocusable, IEditable, IDisableable, ICanBeReadonly, IHasPlaceholder, TimePickerValue, IDropdownToggler } from '@epam/uui';
+import customParseFormat from "dayjs/plugin/customParseFormat";
+import { Dropdown } from '../overlays';
+dayjs.extend(customParseFormat);
+
+export interface BaseTimePickerProps extends IEditable<TimePickerValue>, IDisableable, ICanBeReadonly, IHasPlaceholder {
+    minutesStep?: number;
+    format?: 12 | 24;
+    renderTarget?(props: IDropdownToggler): React.ReactNode;
+}
+
+interface TimePickerState {
+    isOpen: boolean;
+    value: string | null;
+}
+
+const valueToTimeString = (value: TimePickerValue, format: BaseTimePickerProps['format']) => {
+    if (value === null) return null;
+    return dayjs().set(value).format(format === 24 ? 'HH:mm' : 'hh:mm A');
+};
+
+export abstract class BaseTimePicker<TProps extends BaseTimePickerProps> extends React.Component<TProps, TimePickerState> {
+    state = {
+        isOpen: false,
+        value: valueToTimeString(this.props.value, this.props.format),
+    };
+
+    abstract renderInput: (props: IDropdownToggler) => React.ReactNode;
+    abstract renderBody: () => React.ReactNode;
+
+    componentDidUpdate(prevProps: BaseTimePickerProps) {
+        if (this.props.value !== prevProps.value) {
+            this.setState({ ...this.state, value: valueToTimeString(this.props.value, this.props.format) });
+        }
+    }
+
+    getFormat = () => {
+        return this.props.format === 24 ? 'HH:mm' : 'hh:mm A';
+    }
+
+    onClear = () => {
+        this.props.onValueChange({ hours: 0, minutes: 0 });
+    }
+
+    onToggle = (value: boolean) => {
+        this.setState({ ...this.state, isOpen: value });
+    }
+
+    handleInputChange = (newValue: string) => {
+        if (this.getFormat() === "hh:mm A" && newValue.length < 8) {
+            this.setState({ ...this.state, value: newValue });
+        } else if (dayjs(newValue, this.getFormat(), true).isValid()) {
+            const value = dayjs(newValue, this.getFormat(), true);
+            this.props.onValueChange({ hours: value.hour(), minutes: value.minute() });
+            this.setState({ ...this.state, value: newValue });
+        } else {
+            this.setState({ ...this.state, value: newValue });
+        }
+    }
+
+    handleFocus = (e: React.FocusEvent<HTMLInputElement>) => {
+        this.onToggle(true);
+    }
+
+    handleBlur = (e: React.FocusEvent<HTMLInputElement>) => {
+        if (isChildFocusable(e)) return;
+        this.onToggle(false);
+
+        if (this.state.value === '') {
+            this.props.onValueChange(null);
+            this.setState({ ...this.state, value: null });
+        } else if (!dayjs(this.state.value, this.getFormat(), true).isValid()) {
+            this.props.onValueChange(this.props.value);
+            this.setState({ ...this.state, value: valueToTimeString(this.props.value, this.props.format) });
+        }
+    }
+
+    render() {
+        return (
+            <Dropdown
+                renderTarget={ (props: IDropdownToggler) => this.props.renderTarget ? this.props.renderTarget(props) : this.renderInput(props) }
+                renderBody={ () => !this.props.isDisabled && !this.props.isReadonly && this.renderBody() }
+                onValueChange={ !this.props.isDisabled && !this.props.isReadonly ? this.onToggle : null }
+                value={ this.state.isOpen }
+                modifiers={ [{ name: 'offset', options: { offset: [0, 6] } }] }
+            />
+        );
+    }
+}
